Move cluster info service to proxy-nsmd-k8s
